requirements.txt
streamlit
pandas
numpy
tushare
google-genai
plotly